init plutarch-benchmark
